+++ app/src/main/java/com/ml/wting/repository/model/BannerItem.kt
@@ -2,5 +2,5 @@ package com.ml.wting.repository.model
 
 data  class BannerItem(
     val imageUrl:String,
-    val targetId:Int
+    val targetId:Long
 )
