+++ app/src/main/java/com/ml/wting/repository/viewmodel/MVDetailViewModel.kt
@@ -18,7 +18,7 @@ class MVDetailViewModel : BaseViewModel() {
     }
 
 
-    fun getMVDetail(mvid: Int): LiveData<MVDetail> {
+    fun getMVDetail(mvid: Long): LiveData<MVDetail> {
 
 
         request(apiService.getMVDetail(mvid),
